perf(memory): const relationship patterns + early-out in extract_facts
extract_facts runs on every user utterance (auto-capture). Its extract_relationships helper rebuilt 57 format! String patterns on every call (3 phrasings x 19 relations), each a heap allocation, then scanned the text for each — even when the utterance contained no relationship phrase at all.

Replace the per-call format! loop with a const table of &'static str patterns (byte-identical to the old format! output — verified: all 76 literals equal the format! templates in order), and add an early-out: the input is already lowercased at the call site and every pattern starts with "my ", so when the text contains no "my " the whole 19-relation scan is skipped. Matched substrings, the split predicate, and the returned (relation, name) pairs are unchanged.

Measured on an x86 dev box (300k calls/run): the common no-relationship utterance goes 15.87us -> 2.79us per call (~5.7x, early-out skips the scan + 57 allocations); utterances containing "my " go ~15.9us -> ~9.5us (~1.6x, const table removes the 57 allocations). Jetson numbers in the PR body.

The 24 tests only exercise the public extract_facts/extract_and_store API, so they move to tests/extract_test.rs (integration); tests/extract_bench.rs adds an ignored microbench. Contributes under #402 (performance bucket).
